requirements.txt
appdirs==1.4.3
Django==1.11.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
pandas
graphlab-create
geopy
